auth/auth_token.py
from datetime import timedelta, datetime, timezone
from typing import Optional
import schemas
import jwt

from jose import jwt, JWTError

SECRET_KEY = "7AlBaQQSuVfxYVeLyKX1G-4WfVc2PTdxE7BYQIUeSOYmxmRXpo4AqbfkFLLsjR7TZA_7p9Fu38raedB2pH31Og"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now(timezone.utc) + expires_delta
    else:
        expire = datetime.now(timezone.utc) + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, ALGORITHM)
    return encoded_jwt


def verify_token(token: str, credentials_exception):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        email: str = payload.get("sub")
        if email is None:
            raise credentials_exception
        return schemas.TokenData(email=email)
    except JWTError:
        raise credentials_exception
